+++ src/main/java/gdg/pium/domain/image/service/ImageService.java
@@ -1,4 +1,4 @@
-package gdg.pium.image.service;
+package gdg.pium.domain.image.service;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
