+++ src/main/java/gdg/pium/domain/post/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package gdg.pium.post.repository;
+package gdg.pium.domain.post.repository;
 
-import gdg.pium.post.Post;
+import gdg.pium.domain.post.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
