+++ src/main/java/gdg/pium/domain/user/Provider.java
@@ -1,4 +1,4 @@
-package gdg.pium.user;
+package gdg.pium.domain.user;
 
 public enum Provider {
