+++ src/main/java/gdg/pium/domain/user/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package gdg.pium.domain.user.repository;
+
+import gdg.pium.domain.user.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
