--- src/main/java/gdg/pium/user/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package gdg.pium.user.repository;
-
-import gdg.pium.user.Users;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<Users, Long> {
-}
